refactor: getUserInput(), printResult()와 같이 입출력에 관련된 메서드를 CalculatorIO Class 내부로 분리
