+++ src/main/java/calculator/Calculator.java
@@ -1,10 +1,9 @@
-package calculator.domain;
+package calculator;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class Calculator {
-    public String userInput;
     public List<String> processedInput = new ArrayList<>();
     public int total;
 }
